Revert "Simplify reading project string #3406"
This reverts commit c54a2c064017fff643eb2f50c6df6b7242b8e91e.
